Moved main dashboard page under pages/main
