+++ entropylab/results/dashboard/pages/main/auto_plot.py
